requirements.txt
pyserial>=3.5
gradio>=3.4